reorganize project, get tests passing with new structure
