+++ nereid/Dockerfile.multi
@@ -7,6 +7,7 @@
 
 From redis:5.0.7-alpine as redis
 COPY redis.conf /redis.conf
+CMD ["redis-server", "/redis.conf"]
 
 
 FROM python:alpine as flower
