+++ nereid/nereid/api/api_v1/endpoints/network.py
@@ -12,14 +12,15 @@ from starlette.templating import Jinja2Templates
 import celery
 from celery.result import AsyncResult
 
-from ..models.network import Graph, Node, SubgraphRequest
 from nereid.bg_worker import (
     background_validate_network_from_dict,
     background_network_subgraphs,
 )
+
 from nereid.core import config
-from nereid.network.utils import graph_factory
-from nereid.network.render import render, fig_to_image
+from nereid.src.network.utils import graph_factory
+from nereid.src.network.render import render, fig_to_image
+from nereid.src.network.models import Graph, Node, SubgraphRequest
 
 
 router = APIRouter()
