+++ nereid/nereid/bg_worker.py
@@ -1,13 +1,12 @@
 import networkx as nx
 from nereid.core.celery_app import celery_app
 
-from nereid.network.network_validate import validate_network_from_dict
-from nereid.network.network_algo import network_subgraphs
+from nereid.src.network.tasks import network_subgraphs, validate_network
 
 
 @celery_app.task(acks_late=True, track_started=True)
 def background_validate_network_from_dict(graph):
-    return validate_network_from_dict(graph)
+    return validate_network(graph)
 
 
 @celery_app.task(acks_late=True, track_started=True)
